Update html2ors.py
Created a function findinitiallink() to obtain the initial link for the article. This function uses if/else tests to identify the correct way to get the link from the HTML file. Also, the urllib library was imported to change the html coding to allow creation of a popup URL. Further revision needed to complete the popupURL link as more parts of the code were changed in the GS html code.